WIP: Move CommandParser out of core
Removed unused requires in files touched
