+++ bundles/ranvier-lib/lib/CommandType.js
