+++ src/BundleManager.js
@@ -6,7 +6,6 @@ const fs = require('fs'),
     Data = require('./Data'),
     Area = require('./Area'),
     Command = require('./Command'),
-    CommandType = require('./CommandType'),
     Item = require('./Item'),
     Npc = require('./Npc'),
     PlayerClass = require('./PlayerClass'),
